implement hero data endpoints
